[Refactor/#138] 수행한 퀘스트(챌린지) 상세 화면을 screens.challenge_detail 패키지로 이동
