+++ feature/my/src/main/java/com/ilsangtech/ilsang/feature/my/screens/challenge_detail/MyChallengeViewModel.kt
@@ -1,4 +1,4 @@
-package com.ilsangtech.ilsang.feature.my
+package com.ilsangtech.ilsang.feature.my.screens.challenge_detail
 
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
